Address code review: fix redundant toFloat and update splash comment
Agent-Logs-Url: https://github.com/upamanyu92/Emotion-aware-ai-android/sessions/9baf069e-bdcf-49c8-88db-0a54529d8e05

+++ app/src/main/java/com/example/emotionawareai/ui/screen/SplashScreen.kt
@@ -51,7 +51,7 @@ import kotlin.math.sin
  *
  * Features:
  * - Deep gradient background with floating particle animation
- * - Banner image (IMG_0232.PNG) that scales up with spring-like easing
+ * - Banner image (splash_banner) that scales up with spring-like easing
  * - Pulsating glow ring evoking a heartbeat / emotional-awareness theme
  * - Tagline that fades in gracefully
  *
@@ -153,7 +153,7 @@ fun SplashScreen(onFinished: () -> Unit) {
                 drawCircle(
                     color = color.copy(alpha = particleAlpha),
                     radius = particleRadius.dp.toPx(),
-                    center = Offset(cx.toFloat(), cy.toFloat())
+                    center = Offset(cx, cy)
                 )
             }
         }
